requirements.txt
Django==5.0.3
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2024.1
xmltodict==0.13.0
djangorestframework==3.15.1
# Django Composite Foreignkey
django-composite-foreignkey==1.1.0
psycopg2-binary
django-cors-headers
jwcrypto== 1.5.6
pycparser== 2.22